+++ tests/Unit/Requests/GetParcelListTest.php
@@ -45,7 +45,7 @@ class GetParcelListTest extends TestCase
     }
 
     /** @test */
-    public function it_can_return_a_reponse()
+    public function it_can_return_a_response()
     {
         $request = new GetParcelList;