+++ tests/Unit/Requests/ModifyCODTest.php
@@ -20,7 +20,7 @@ class ModifyCODTest extends TestCase
     }
 
     /** @test */
-    public function it_can_return_a_reponse()
+    public function it_can_return_a_response()
     {
         $request = new ModifyCod(1, 2, 10);