+++ tests/Unit/Requests/PrepareLabelsTest.php
@@ -21,7 +21,7 @@ class PrepareLabelsTest extends TestCase
     }
 
     /** @test */
-    public function it_can_return_a_reponse()
+    public function it_can_return_a_response()
     {
         $request = new PrepareLabels;