+++ tests/Unit/Requests/PrintLabelsTest.php
@@ -55,7 +55,21 @@ class PrintLabelsTest extends TestCase
     }
 
     /** @test */
-    public function it_can_return_a_reponse()
+    public function it_can_set_the_type_of_printer()
+    {
+        $request = new PrintLabels;
+        $request->typeOfPrinter('Thermo');
+
+        $this->assertEquals([
+            'ParcelList' => [],
+            'PrintPosition' => 1,
+            'ShowPrintDialog' => false,
+            'TypeOfPrinter' => 'Thermo',
+        ], $request->toArray());
+    }
+
+    /** @test */
+    public function it_can_return_a_response()
     {
         $request = new PrintLabels;